Refactored code, with new class handling list callbacks
